requirement.txt
pdfminer3k
Pyocr